Better fix for connected test - shutdown the old stairway
